Add `Setup Chromedriver` action
Add the `Setup Chromedriver` action, which is the core of this project.

This action consolidates the process of downloading the
`chromium-browser`, `chromium-chromedriver`, and `chromium-codes-*`
files from Ubuntu's pool archive, and installing them.

This is needed over simply using the repository, because starting with
Ubuntu 19.10 the `chromium-browser` deb was replaced with a transitional
package redirecting to the snap. The problem is that snaps won't install
in an Ubuntu Docker image.